refactor: dto의 인덱스 관련 데이터 타입을 int로 변경

+++ backend/src/main/java/reviewme/highlight/service/HighlightService.java
@@ -55,7 +55,7 @@ public class HighlightService {
             for (HighlightedLineRequest line : highlight.lines()) {
                 for (HighlightIndexRangeRequest range : line.ranges()) {
                     Highlight highLight = new Highlight(highlight.answerId(),
-                            line.index().intValue(), range.startIndex().intValue(), range.endIndex().intValue());
+                            line.index(), range.startIndex(), range.endIndex());
                     highlights.add(highLight);
                 }
             }
