+++ backend/src/main/java/reviewme/highlight/service/dto/HighlightIndexRangeRequest.java
@@ -5,9 +5,9 @@ import jakarta.validation.constraints.NotNull;
 public record HighlightIndexRangeRequest(
 
         @NotNull(message = "시작 인덱스를 입력해주세요.")
-        Long startIndex,
+        int startIndex,
 
         @NotNull(message = "끝 인덱스를 입력해주세요.")
-        Long endIndex
+        int endIndex
 ) {
 }
